@@ -1,18 +1,17 @@
 [build-system]
-requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
-build-backend = "setuptools.build_meta"
+requires = ["setuptools>=60", "setuptools-scm>=8.0"]
 
 [tool.setuptools_scm]
-write_to = "_version.py"
+version_file="_version.py"
 
 [tool.black]
 line-length = 95
-target-version = ["py310"]
+target-version = ["py312"]
 extend-exclude = '''^(.*\/)*\b(migrations)\b($|\/.*$)'''
 
 [tool.isort]
 profile = "black"
-py_version = "310"
+py_version = "312"
 skip = [".tox", ".eggs", "migrations"]
 
 [tool.coverage.run]
@@ -31,3 +30,9 @@ exclude_lines = [
     "pragma: no cover",
     "if TYPE_CHECKING:",
 ]
+
+[tool.pytest.ini_options]
+django_find_project = false
+DJANGO_SETTINGS_MODULE = "multisite.tests.test_settings"
+python_files = "tests.py test_*.py *_tests.pyc"
+addopts = "--reuse-db"
